+++ src/docs/pages/component/link/index.vue
@@ -27,10 +27,10 @@
     |类型|支持|默认|
     |-|-|-|
     |尺寸|全部|`m`|
-    |色彩|全部|`theme`|
+    |色彩|全部(除`transparent`)|`theme`|
     |状态|全部|`normal`|
 
-    <a href="/guide/status.html">查看形态文档</a>
+    <a href="/guide/status.html">查看形态文档</a> / <a href="/design/color.html">查看色彩文档</a>
 
     :::vue
     @name:尺寸