+++ src/docs/pages/component/quote/index.vue
@@ -27,10 +27,10 @@
     |类型|支持|默认|
     |-|-|-|
     |尺寸|不支持|-|
-    |色彩|全部|`light-blue`|
+    |色彩|全部(除`transparent`)|`neutral-6`|
     |状态|不支持|-|
 
-    <a href="/guide/status.html">查看形态文档</a>
+    <a href="/guide/status.html">查看形态文档</a> / <a href="/design/color.html">查看色彩文档</a>
 
     :::vue
     @layout:color