+++ src/docs/pages/component/table/index.vue
@@ -365,10 +365,10 @@
     |类型|支持|默认|
     |-|-|-|
     |尺寸|`xs` / `s` / `m` / `l` / `xl`|`m`|
-    |色彩|全部|`light-gray`|
+    |色彩|全部(除`transparent`)|`neutral-2`|
     |状态|不支持|-|
 
-    <a href="/guide/status.html">查看形态文档</a>
+    <a href="/guide/status.html">查看形态文档</a> / <a href="/design/color.html">查看色彩文档</a>
 
     :::vue
     @name:尺寸
@@ -835,7 +835,7 @@
     |width|列的宽度，此配置会改变列在宽度分配时的权重，但不一定完全符合设置的宽度值，如果需要最小最大宽度使用`minwidth`和`maxwidth`。<br>如果需要设置标题列的宽度请使用`title-col-width`配置。|宽度数值(必需包含单位`px`或`%`)|String|`undefined`|
     |minwidth|列的最小宽度，注意列会尽可能遵循此值，但不是绝对一致的。|宽度数值(必需包含单位`px`或`%`)|String|`undefined`|
     |maxwidth|列的最小宽度，注意列会尽可能遵循此值，但不是绝对一致的。|宽度数值(必需包含单位`px`或`%`)|String|`undefined`|
-    |style|设置列的色彩样式|[形态/颜色](/guide/status.html#颜色)中所有的颜色值|String|`undefined`|
+    |style|设置列的色彩样式|[设计/色彩](/design/color.html)中所有的颜色值|String|`undefined`|
     |disabled|是否展示列的禁用样式|`true`<br>`false`|Boolean|`false`|
     |align|列的对齐方式|`'left'`<br>`'right'`<br>`'align'`|String|`'align'`|
     |title|此列是否是标题列，根据`fixed-title-col`的设置所有的标题列会被放到最左或最右|`true`<br>`false`|Boolean|`false`|
@@ -1147,7 +1147,7 @@
     |KEY|描述|接受值|值类型|默认值|
     |-|-|-|-|-|
     |row|需要设置的行数(从0开始，在有标题行时，0表示标题行)|行数|Number|`undefined`|
-    |style|设置行的色彩样式|[形态/颜色](/guide/status.html#颜色)中所有的颜色值|String|`undefined`|
+    |style|设置行的色彩样式|[设计/色彩](/design/color.html)中所有的颜色值|String|`undefined`|
     |disabled|是否展示行的禁用样式|`true`<br>`false`|Boolean|`false`|
     |align|行的对齐方式|`'left'`<br>`'right'`<br>`'align'`|String|`'align'`|
     |disableSelection|此行在多选模式下处于禁用状态，用户无法改变选中状态，但`setCheckedRows()`方法可以改变选中状态（必须配合`multi-select`才生效）|`true`<br>`false`|Boolean|`false`|
@@ -1265,7 +1265,7 @@
     |-|-|-|-|-|
     |row|需要设置单元格的行数(从0开始，在有标题行时，0表示标题行)|行数|Number|`undefined`|
     |col|需要设置单元格所在列的KEY|列的KEY(`list`配置中设置)|String|`undefined`|
-    |style|设置单元格的色彩样式|[形态/颜色](/guide/status.html#颜色)中所有的颜色值|String|`undefined`|
+    |style|设置单元格的色彩样式|[设计/色彩](/design/color.html)中所有的颜色值|String|`undefined`|
     |disabled|是否展示单元格的禁用样式|`true`<br>`false`|Boolean|`false`|
     |align|单元格的对齐方式|`'left'`<br>`'right'`<br>`'align'`|String|`'align'`|
     <br>