+++ src/docs/pages/component/textinput/index.vue
@@ -54,7 +54,7 @@
     |色彩|不支持|-|
     |状态|`normal`<br/>`disabled`<br>`readonly`|`normal`|
 
-    <a href="/guide/status.html">查看形态文档</a>
+    <a href="/guide/status.html">查看形态文档</a> / <a href="/design/color.html">查看色彩文档</a>
 
     :::vue
     @name:尺寸