+++ src/docs/pages/guide/logs/index.vue
@@ -12,15 +12,61 @@
 
     如果更新对使用方法有影响，会在更新记录的`Upgrade Precautions` / `升级注意事项`中说明。
 
-    <!-- ### 1.0.0 <ui-badge class="circle" s light-black>Stable</ui-badge> -->
+    <!-- ### 1.0.0 <ui-badge class="circle" s neutral-10>Stable</ui-badge> -->
 
-    <!-- ### 0.11.0 <ui-badge class="circle" s light-blue>Beta</ui-badge> -->
+    <!-- ### 0.11.0 <ui-badge class="circle" s neutral-8>Beta</ui-badge> -->
     
     ### 更新记录
 
+    ### 0.12.34 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
+    
+    <ui-textcolor color="light-blue">
+        <ui-small>
+            发布于2019年5月20日，
+            Archive :
+            <a href="https://github.com/Morning-UI/morning-ui/releases/tag/0.12.34" target="_blank">Package</a>
+            <a href="https://cdn.jsdelivr.net/npm/morning-ui@0.12.34/dist/" target="_blank">jsdelivr(CDN)</a>
+            <a href="https://unpkg.com/morning-ui@0.12.34/dist/" target="_blank">unpkg(CDN)</a>
+        </ui-small>
+    </ui-textcolor>
+    
+    #### 改进
+    
+    - `辅助色`现在被改进为`中性色`，原有的12种辅助色现在调整为14种中性色，详见文档的`设计/色彩`章节
+    - 所有组件中使用和支持的`辅助色`现在被调整为`中性色`
+    - 优化`ui-select`、`ui-multiinput`、`ui-multiform`中的多选项目样式 
+    
+    #### 修复
+    
+    - 修复`ui-multiinput`可以输入空数值的问题
+    - 修复`ui-steps`加载中状态未显示的问题
+    
+    #### 更多
+    
+    - 新增`设计/色彩`文档，原有的`形态/色彩`部分内容被调整至此文档
+    - `ui-menu`移除4种色彩支持：`gray`、`light-gray`、`light-black`、`extra-light-black`
+    
+    <ui-label color="primary">升级注意事项</ui-label>
+    
+    - `ui-menu`移除4种色彩支持，请替换为`white`或`black`
+    - `辅助色`现在被改进为`中性色`，在使用`color`配置以及配色时需要注意以下替换（色值不一致）：
+        - `light-black` 替换为 `neutral-11`
+        - `extra-light-black` 替换为 `neutral-10`
+        - `blue` 替换为 `neutral-9`
+        - `light-blue` 替换为 `neutral-8`
+        - `extra-light-blue` 替换为 `neutral-7`
+        - `silver` 替换为 `neutral-6`
+        - `light-silver` 替换为 `neutral-5`
+        - `extra-light-silver` 替换为 `neutral-4`
+        - `light-gray` 替换为 `neutral-3`
+        - `gray` 替换为 `neutral-2`
+        - 新增`neutral-1`、`transparent`
+    
+    ---
+
     ### 0.12.33 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年5月16日，
             Archive :
@@ -48,7 +94,7 @@
 
     ### 0.12.32 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年5月13日，
             Archive :
@@ -64,7 +110,7 @@
 
     ### 0.12.31 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年5月10日，
             Archive :
@@ -80,7 +126,7 @@
 
     ### 0.12.30 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年5月6日，
             Archive :
@@ -102,7 +148,7 @@
 
     ### 0.12.29 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年5月6日，
             Archive :
@@ -121,7 +167,7 @@
 
     ### 0.12.28 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月25日，
             Archive :
@@ -140,7 +186,7 @@
 
     ### 0.12.27 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月23日，
             Archive :
@@ -163,7 +209,7 @@
 
     ### 0.12.26 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月23日，
             Archive :
@@ -198,7 +244,7 @@
 
     ### 0.12.25 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月21日，
             Archive :
@@ -234,7 +280,7 @@
 
     ### 0.12.24 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月20日，
             Archive :
@@ -265,7 +311,7 @@
 
     ### 0.12.23 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月19日，
             Archive :
@@ -284,7 +330,7 @@
 
     ### 0.12.22 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月19日，
             Archive :
@@ -319,7 +365,7 @@
 
     ### 0.12.21 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月14日，
             Archive :
@@ -344,7 +390,7 @@
 
     ### 0.12.20 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月12日，
             Archive :
@@ -367,7 +413,7 @@
 
     ### 0.12.19 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月11日，
             Archive :
@@ -390,7 +436,7 @@
 
     ### 0.12.18 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月10日，
             Archive :
@@ -408,7 +454,7 @@
 
     ### 0.12.17 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月7日，
             Archive :
@@ -451,7 +497,7 @@
 
     ### 0.12.16 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月4日，
             Archive :
@@ -482,7 +528,7 @@
 
     ### 0.12.15 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年4月3日，
             Archive :
@@ -509,7 +555,7 @@
 
     ### 0.12.14 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月29日，
             Archive :
@@ -527,7 +573,7 @@
 
     ### 0.12.13 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月28日，
             Archive :
@@ -545,7 +591,7 @@
     
     ### 0.12.12 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月25日，
             Archive :
@@ -572,7 +618,7 @@
 
     ### 0.12.11 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月22日，
             Archive :
@@ -594,7 +640,7 @@
     
     ### 0.12.10 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月21日，
             Archive :
@@ -621,7 +667,7 @@
 
     ### 0.12.9 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月17日，
             Archive :
@@ -652,7 +698,7 @@
 
     ### 0.12.8 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月15日，
             Archive :
@@ -691,7 +737,7 @@
 
     ### 0.12.7 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月12日，
             Archive :
@@ -726,7 +772,7 @@
 
     ### 0.12.6 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月10日，
             Archive :
@@ -785,7 +831,7 @@
 
     ### 0.12.5 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月4日，
             Archive :
@@ -816,7 +862,7 @@
     
     ### 0.12.4 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年3月1日，
             Archive :
@@ -842,7 +888,7 @@
 
     ### 0.12.3 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年2月26日，
             Archive :
@@ -887,7 +933,7 @@
     
     ### 0.12.2 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年2月22日，
             Archive :
@@ -927,7 +973,7 @@
 
     ### 0.12.1 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年2月19日，
             Archive :
@@ -957,7 +1003,7 @@
 
     ### 0.12.0 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年2月8日，
             Archive :
@@ -1030,9 +1076,9 @@
         
     ---
 
-    ### 0.11.17 <ui-badge class="circle ver-beta" size="s" color="extra-light-black">Beta</ui-badge>
+    ### 0.11.17 <ui-badge class="circle ver-beta" size="s" color="neutral-9">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2019年1月3日，
             Archive :
@@ -1054,9 +1100,9 @@
 
     ---
 
-    ### 0.11.16 <ui-badge class="circle ver-beta" size="s" color="extra-light-black">Beta</ui-badge>
+    ### 0.11.16 <ui-badge class="circle ver-beta" size="s" color="neutral-9">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年11月14日，
             Archive :
@@ -1082,9 +1128,9 @@
 
     ---
 
-    ### 0.11.15 <ui-badge class="circle ver-beta" size="s" color="extra-light-black">Beta</ui-badge>
+    ### 0.11.15 <ui-badge class="circle ver-beta" size="s" color="neutral-9">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年8月28日，
             Archive :
@@ -1107,7 +1153,7 @@
 
     ### 0.11.14 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年8月25日，
             Archive :
@@ -1143,7 +1189,7 @@
 
     ### 0.11.13 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年8月22日，
             Archive :
@@ -1183,9 +1229,9 @@
 
     ---
 
-    ### 0.11.12 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.11.12 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年8月19日，
             Archive :
@@ -1220,7 +1266,7 @@
 
     ### 0.11.11 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年8月15日，
             Archive :
@@ -1257,7 +1303,7 @@
 
     ### 0.11.10 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月26日，
             Archive :
@@ -1285,7 +1331,7 @@
 
     ### 0.11.9 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月23日，
             Archive :
@@ -1312,7 +1358,7 @@
 
     ### 0.11.8 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月16日，
             Archive :
@@ -1347,7 +1393,7 @@
 
     ### 0.11.7 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月10日，
             Archive :
@@ -1374,7 +1420,7 @@
 
     ### 0.11.6 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月4日，
             Archive :
@@ -1398,7 +1444,7 @@
 
     ### 0.11.5 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月3日，
             Archive :
@@ -1422,7 +1468,7 @@
 
     ### 0.11.4 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年7月1日，
             Archive :
@@ -1453,7 +1499,7 @@
 
     ### 0.11.3 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年6月28日，
             Archive :
@@ -1481,7 +1527,7 @@
 
     ### 0.11.2 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年6月25日，
             Archive :
@@ -1508,7 +1554,7 @@
 
     ### 0.11.1 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年6月24日，
             Archive :
@@ -1536,7 +1582,7 @@
 
     ### 0.11.0 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年6月23日，
             Archive :
@@ -1617,9 +1663,9 @@
 
     ---
 
-    ### 0.10.32 <ui-badge class="circle ver-stable" size="s" color="extra-light-black">Stable</ui-badge>
+    ### 0.10.32 <ui-badge class="circle ver-stable" size="s" color="neutral-9">Stable</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年6月5日，
             Archive : 
@@ -1636,9 +1682,9 @@
 
     ---
 
-    ### 0.10.31 <ui-badge class="circle ver-stable" size="s" color="extra-light-black">Stable</ui-badge>
+    ### 0.10.31 <ui-badge class="circle ver-stable" size="s" color="neutral-9">Stable</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月28日，
             Archive : 
@@ -1655,9 +1701,9 @@
 
     ---
 
-    ### 0.10.30 <ui-badge class="circle ver-stable" size="s" color="extra-light-black">Stable</ui-badge>
+    ### 0.10.30 <ui-badge class="circle ver-stable" size="s" color="neutral-9">Stable</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月24日，
             Archive : 
@@ -1674,9 +1720,9 @@
 
     ---
 
-    ### 0.10.29 <ui-badge class="circle ver-stable" size="s" color="extra-light-black">Stable</ui-badge>
+    ### 0.10.29 <ui-badge class="circle ver-stable" size="s" color="neutral-9">Stable</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月22日，
             Archive : 
@@ -1691,9 +1737,9 @@
 
     ---
 
-    ### 0.10.28 <ui-badge class="circle ver-stable" size="s" color="extra-light-black">Stable</ui-badge>
+    ### 0.10.28 <ui-badge class="circle ver-stable" size="s" color="neutral-9">Stable</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月22日，
             Archive : 
@@ -1713,9 +1759,9 @@
 
     ---
 
-    ### 0.10.27 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.10.27 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月20日，
             Archive : 
@@ -1736,9 +1782,9 @@
 
     ---
 
-    ### 0.10.26 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.10.26 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月18日，
             Archive : 
@@ -1760,7 +1806,7 @@
 
     ### 0.10.25 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月17日，
             Archive : 
@@ -1790,9 +1836,9 @@
 
     ---
 
-    ### 0.10.24 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.10.24 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月14日，
             Archive : 
@@ -1814,9 +1860,9 @@
 
     ---
 
-    ### 0.10.23 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.10.23 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
     
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月12日，
             Archive : 
@@ -1833,9 +1879,9 @@
 
     ---
 
-    ### 0.10.22 <ui-badge class="circle ver-beta" size="s" color="light-blue">Beta</ui-badge>
+    ### 0.10.22 <ui-badge class="circle ver-beta" size="s" color="neutral-8">Beta</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年5月9日，
             Archive : 
@@ -1854,7 +1900,7 @@
 
     ### 0.10.21 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年4月22日，
             Archive : 
@@ -1873,7 +1919,7 @@
 
     ### 0.10.20 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年4月18日，
             Archive : 
@@ -1892,7 +1938,7 @@
 
     ### 0.10.19 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年4月15日，
             Archive : 
@@ -1914,7 +1960,7 @@
 
     ### 0.10.18 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年4月13日，
             Archive : 
@@ -1937,7 +1983,7 @@
 
     ### 0.10.17 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年2月26日，
             Archive : 
@@ -1955,7 +2001,7 @@
 
     ### 0.10.16 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年1月31日，
             Archive : 
@@ -1972,7 +2018,7 @@
 
     ### 0.10.15 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年1月16日，
             Archive : 
@@ -1990,7 +2036,7 @@
         
     ### 0.10.14 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2018年1月15日，
             Archive : 
@@ -2014,7 +2060,7 @@
 
     ### 0.10.13 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年12月7日，
             Archive : 
@@ -2034,7 +2080,7 @@
 
     ### 0.10.12 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年12月5日，
             Archive : 
@@ -2054,7 +2100,7 @@
 
     ### 0.10.11 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月29日，
             Archive : 
@@ -2079,7 +2125,7 @@
 
     ### 0.10.10 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月27日，
             Archive : 
@@ -2102,7 +2148,7 @@
 
     ### 0.10.9 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月23日，
             Archive : 
@@ -2121,7 +2167,7 @@
 
     ### 0.10.8 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月20日，
             Archive : 
@@ -2152,7 +2198,7 @@
 
     ### 0.10.7 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月13日，
             Archive : 
@@ -2173,7 +2219,7 @@
 
     ### 0.10.6 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月20日，
             Archive : 
@@ -2204,7 +2250,7 @@
 
     ### 0.10.7 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月13日，
             Archive : 
@@ -2225,7 +2271,7 @@
 
     ### 0.10.6 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月10日，
             Archive : 
@@ -2248,7 +2294,7 @@
 
     ### 0.10.5 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
             发布于2017年11月6日，
             Archive : 
@@ -2266,7 +2312,7 @@
 
     ### 0.10.4 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
         发布于2017年11月5日，
             Archive : 
@@ -2290,7 +2336,7 @@
 
     ### 0.10.3 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
         发布于2017年11月2日，
             Archive : 
@@ -2319,7 +2365,7 @@
 
     ### 0.10.2 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
         发布于2017年11月1日，
             Archive : 
@@ -2338,7 +2384,7 @@
 
     ### 0.10.1 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
         发布于2017年10月31日，
             Archive : 
@@ -2360,7 +2406,7 @@
 
     ### 0.10.0 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
-    <ui-textcolor color="light-blue">
+    <ui-textcolor color="neutral-8">
         <ui-small>
         发布于2017年10月31日，
             Archive :