+++ src/lib/components/multiinput/index.vue
@@ -32,7 +32,7 @@
         >
             <span :title="value">{{value}}</span>
             <i
-                class="mo-icon mo-icon-error-cf"
+                class="mo-icon mo-icon-close"
                 v-if="conf.state !== 'disabled' && conf.state !== 'readonly'"
                 @click="_deleteItem(index)"
             ></i>
@@ -222,14 +222,18 @@ export default {
 
                 let $input = this.$el.querySelector('input');
 
-                this.add($input.value);
-                this._blurInput();
+                if ($input.value) {
 
-                setTimeout(() => {
+                    this.add($input.value);
+                    this._blurInput();
 
-                    this._focusInput();
+                    setTimeout(() => {
 
-                });
+                        this._focusInput();
+
+                    });
+
+                }
 
             }